:recycle: Moved style-fn to utils file
