+++ src/molecules/AnimatedCheckmark/AnimatedCheckmark.styles.ts
@@ -1,6 +1,7 @@
 import { tokens } from '@equinor/eds-tokens';
 
 import { AnimatedCheckmarkProps } from './AnimatedCheckmark';
+import { sizeToPx } from './AnimatedCheckmark.utils';
 
 import styled, { keyframes } from 'styled-components';
 
@@ -19,16 +20,6 @@ const animateCheckmark = keyframes`
 
 interface ContainerProps {
   $size: AnimatedCheckmarkProps['size'];
-}
-
-export function sizeToPx(size: AnimatedCheckmarkProps['size']): string {
-  switch (size) {
-    case 'small':
-      return '70px';
-    default:
-    case 'medium':
-      return '144px';
-  }
 }
 
 export const Container = styled.svg<ContainerProps>`
